Identity Server

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts.DTOs;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -48,12 +49,13 @@ namespace AuctionService.Controllers
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto) 
         {
             var auction = _mapper.Map<Auction>(auctionDto);
-            // TODO: add current user as seller
-            auction.Seller = "test";
+            
+            auction.Seller = User.Identity.Name;
 
             _context.Auctions.Add(auction);
 
@@ -71,6 +73,7 @@ namespace AuctionService.Controllers
                 newAuction);
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto) 
         {
@@ -78,7 +81,7 @@ namespace AuctionService.Controllers
 
             if (auction == null) return NotFound();
             
-            // TODO: check seller == username
+            if (auction.Seller != User.Identity.Name) return Forbid("");
 
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -95,6 +98,7 @@ namespace AuctionService.Controllers
             return BadRequest("Problem saving changes");
         }
 
+        [Authorize] 
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id) 
         {
@@ -102,11 +106,12 @@ namespace AuctionService.Controllers
 
             if (auction == null) return NotFound();
             
-            // TODO: check seller == username
+            if (auction.Seller != User.Identity.Name) return Forbid();
 
             _context.Auctions.Remove(auction);
 
-            await _publishEndpoint.Publish(new { Id = auction.Id.ToString() });
+            // await _publishEndpoint.Publish(new { Id = auction.Id.ToString() });
+            await _publishEndpoint.Publish(new AuctionDeleted { Id = auction.Id.ToString() });
 
             var result = await _context.SaveChangesAsync() > 0;
